feat: Create Server class
The Server call extends Thread class so it can start a thread when we start the server in the MainServerThread. It has a list of dedicated servers which are threads as well and are started when a client petition has been accepted by the server through the server socket.

Implement startServer method and stopServer method. These methods put the boolean attribute running as true or false. Instead of writing while(true), we have the option to control the execution of the thread.

When we execute accept(), it will create a socket specifically for that client and it will create a dedicated server. Directly, it starts the thread for that client.