feat: Create DedicatedServer class
Once the connection has been established, we can send and receive data and objects through the dedicated servers. The dedicated server contains a list of other dedicated servers to have the option to send the notifications and updates to all other clients when we receive something from one client. This is also a subclass of the Thread class because we start a new thread for each client.