feat: Create ServerCommunication class
This class is a subclass of the Thread class. There will be a thread to manage the connection to the server by creating a socket to the server using the configurations of the server. Implement the startServerCommunication and stopServerCommunication server, which basically starts and stops the communication. We connect or disconnect from the server.