feat: Finalize the implementation of Server-Client connection
Change the DataInputStream and DataOutPutStream to ObjectInputStream and ObjectOutputStream because we are sending objects through the sockets.

The server will be receiving the product and the operation to be done. Meanwhile, the client will always be listening for the model. More like the state of the model. If something has changed, then the list of products will be updated as well as the list of products for the drop-down list

Use the refrescaLlistaCompra method in the view to refresh the list with the new information received from the server.